0x0C. C - More malloc, free

+++ 0x0C-more_malloc_free/2-calloc.c
@@ -16,7 +16,7 @@ if ((nmemb == 0) || (size == 0))
 s = malloc(p);
 if (s == NULL)
 	return (NULL);
-for (unsigned int i = 0; i < p; i++)
+for (unsigned long i = 0; i < p; i++)
 	s[i] = '0';
 return (s);
 }
